requirements.txt
flask
openai
python-dotenv
PyMuPDF
google-auth
google-auth-oauthlib
requests
pywebview
Pillow
